Restructure tooltips into arc labels
The dating reason for a particular relation is now displayed above the
corresponding arc instead of overlaid as a text box.
The new labels are <textPath> elements and part of the svg instead of
regular DOM elements overlaid onto the svg. This fixes several problems,
mainly panning and zooming concerns, which are now handled
automatically by the textPath element.
Some styling work remains.